refractor: saved reducer function and other utilities in separate files
